transpile with typings
