+++ .idea/typescript-compiler.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="TypeScriptCompiler">
-    <option name="isCompilerEnabled" value="true" />
-    <option name="useConfig" value="true" />
+    <option name="enableServiceForProjectsWithoutConfig" value="true" />
+    <option name="recompileOnChanges" value="true" />
     <option name="showAllErrors" value="true" />
   </component>
 </project>
