+++ src/Resources/CoreMixins/ClientAccessor.ts
@@ -1,12 +1,12 @@
 import {nonenumerable} from 'core-decorators';
-import IHydraClient from '../../interfaces';
+import {IHydraClient} from '../../alcaeus';
 import {Constructor} from '../Mixin';
 
 export default function(alcaeus: IHydraClient) {
     return <TBase extends Constructor>(Base: TBase) => {
         class HydraResource extends Base {
             @nonenumerable
-            private get _alcaeus() {
+            public get _alcaeus() {
                 return alcaeus;
             }
         }
