+++ src/Resources/Mixins/DocumentedResource.ts
@@ -1,6 +1,6 @@
 import {Core} from '../../Constants';
-import {IDocumentedResource} from '../../interfaces';
 import {rdfs, Schema} from '../../Vocabs';
+import {IDocumentedResource} from '../index';
 import {Constructor} from '../Mixin';
 
 function getTitle(res) {
