+++ src/Resources/Mixins/RdfProperty.ts
@@ -1,7 +1,8 @@
 import {Core} from '../../Constants';
-import {IRdfProperty, IResource} from '../../interfaces';
 import {rdf, rdfs} from '../../Vocabs';
+import {IRdfProperty} from '../index';
 import {Constructor} from '../Mixin';
+import {IResource} from '../Resource';
 
 export function Mixin<TBase extends Constructor>(Base: TBase) {
     abstract class RdfProperty extends Base implements IRdfProperty {
