+++ src/Resources/Mixins/StatusCodeDescription.ts
@@ -1,6 +1,7 @@
 import {Core} from '../../Constants';
-import {IResource, IStatusCodeDescription} from '../../interfaces';
+import {IStatusCodeDescription} from '../index';
 import {Constructor} from '../Mixin';
+import {IResource} from '../Resource';
 
 export function Mixin <TBase extends Constructor>(Base: TBase) {
     return class StatusCodeDescription extends Base implements IStatusCodeDescription {
