+++ src/Resources/Resource.ts
@@ -1,8 +1,12 @@
 import {nonenumerable} from 'core-decorators';
 import {promises as jsonld} from 'jsonld';
 import {Core, JsonLd} from '../Constants';
-import {IResource} from '../interfaces';
-import TypeCollection from '../TypeCollection';
+import TypeCollection, {ITypeCollection} from '../TypeCollection';
+
+export interface IResource {
+    id: string;
+    types: ITypeCollection;
+}
 
 const isProcessed = new WeakMap<IResource, boolean>();
 
@@ -36,11 +40,11 @@ export default class implements IResource {
         return jsonld.compact(this, context || Core.Context);
     }
 
-    protected _get(property: string) {
+    public _get(property: string) {
         return this[property] || null;
     }
 
-    protected _getArray(property: string) {
+    public _getArray(property: string) {
         const values = this[property];
 
         if (!values) {
