+++ src/internals.ts
@@ -1,4 +1,4 @@
-import {HydraResource} from './interfaces';
+import {HydraResource} from './Resources';
 
 export interface IIncomingLink {
     subjectId: string;
