+++ tsconfig.json
@@ -10,6 +10,7 @@
     "removeComments": false,
     "noImplicitAny": false,
     "suppressImplicitAnyIndexErrors": false,
+    "declaration": true,
     "lib": [
       "es2017",
       "es2015",
